+++ mountainsort5/core/get_times_labels_from_sorting.py
@@ -4,7 +4,7 @@ import numpy.typing as npt
 import spikeinterface as si
 
 
-def get_times_labels_from_sorting(sorting: si.BaseSorting) -> Tuple[npt.NDArray[np.int64], npt.NDArray[np.int32]]:
+def get_times_labels_from_sorting(sorting: si.BaseSorting) -> Tuple[npt.NDArray[np.int64], npt.NDArray[np.intp]]:
     """Get times and labels from a sorting object
     Inputs:
         sorting: a sorting object
